% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod.cure.R
\name{cure_prev}
\alias{cure_prev}
\title{Curing Module}
\usage{
cure_prev(dat, at)
}
\arguments{
\item{dat}{Master data list object of class \code{dat} containing networks,
individual-level attributes, and summary statistics.}

\item{at}{Current time step.}
}
\value{
This function returns \code{dat} after updating the nodal attribute
\code{diag.status} and \code{status}, \code{inf.time}, \code{stage} \code{stage.time} \code{aids.time} \code{diag.stage}
\code{vl} \code{vl.last.usupp} \code{vl.last.supp} \code{diag.time}  \code{last.neg.test} \code{tx.status} \code{cuml.time.on.tx}
\code{cuml.time.off.tx}  \code{tx.period.first}   \code{tx.period.last}    \code{tx.init.time}      \code{count.trans} .
}
\description{
Module for curing at set timepoints to hit prevalence.
}
\keyword{module}
